+++ src/node.rs
@@ -134,15 +134,15 @@ impl Node {
                             let player_update_event: events::PlayerUpdate =
                                 serde_json::from_str(&x).unwrap();
 
-                            let player = lavalink_client
-                                .get_player_context(player_update_event.guild_id)
-                                .unwrap();
-
-                            if let Err(why) = player.update_state(player_update_event.state) {
-                                error!(
-                                    "Error updating state for player {}: {}",
-                                    player_update_event.guild_id.0, why
-                                );
+                            if let Some(player) =
+                                lavalink_client.get_player_context(player_update_event.guild_id)
+                            {
+                                if let Err(why) = player.update_state(player_update_event.state) {
+                                    error!(
+                                        "Error updating state for player {}: {}",
+                                        player_update_event.guild_id.0, why
+                                    );
+                                }
                             }
 
                             ed.parse_and_dispatch(&x, |e| e.player_update).await
@@ -152,15 +152,17 @@ impl Node {
                             "TrackStartEvent" => {
                                 let track_event: events::TrackStart =
                                     serde_json::from_str(&x).unwrap();
-                                let player = lavalink_client
-                                    .get_player_context(track_event.guild_id)
-                                    .unwrap();
 
-                                if let Err(why) = player.update_track(track_event.track.into()) {
-                                    error!(
-                                        "Error sending update track message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
+                                if let Some(player) =
+                                    lavalink_client.get_player_context(track_event.guild_id)
+                                {
+                                    if let Err(why) = player.update_track(track_event.track.into())
+                                    {
+                                        error!(
+                                            "Error sending update track message for player {}: {}",
+                                            track_event.guild_id.0, why
+                                        );
+                                    }
                                 }
 
                                 ed.parse_and_dispatch(&x, |e| e.track_start).await
@@ -168,74 +170,32 @@ impl Node {
                             "TrackEndEvent" => {
                                 let track_event: events::TrackEnd =
                                     serde_json::from_str(&x).unwrap();
-                                let player = lavalink_client
-                                    .get_player_context(track_event.guild_id)
-                                    .unwrap();
 
-                                if let Err(why) = player.finish(track_event.reason.into()) {
-                                    error!(
-                                        "Error sending finish message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
-                                }
+                                if let Some(player) =
+                                    lavalink_client.get_player_context(track_event.guild_id)
+                                {
+                                    if let Err(why) = player.finish(track_event.reason.into()) {
+                                        error!(
+                                            "Error sending finish message for player {}: {}",
+                                            track_event.guild_id.0, why
+                                        );
+                                    }
 
-                                if let Err(why) = player.update_track(track_event.track.into()) {
-                                    error!(
-                                        "Error sending update track message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
+                                    if let Err(why) = player.update_track(track_event.track.into())
+                                    {
+                                        error!(
+                                            "Error sending update track message for player {}: {}",
+                                            track_event.guild_id.0, why
+                                        );
+                                    }
                                 }
 
                                 ed.parse_and_dispatch(&x, |e| e.track_end).await
                             }
                             "TrackExceptionEvent" => {
-                                let track_event: events::TrackException =
-                                    serde_json::from_str(&x).unwrap();
-
-                                let player = lavalink_client
-                                    .get_player_context(track_event.guild_id)
-                                    .unwrap();
-
-                                if let Err(why) = player.finish(true) {
-                                    error!(
-                                        "Error sending finish message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
-                                }
-
-                                if let Err(why) = player.update_track(track_event.track.into()) {
-                                    error!(
-                                        "Error sending update track message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
-                                }
-
                                 ed.parse_and_dispatch(&x, |e| e.track_exception).await
                             }
-                            "TrackStuckEvent" => {
-                                let track_event: events::TrackStuck =
-                                    serde_json::from_str(&x).unwrap();
-
-                                let player = lavalink_client
-                                    .get_player_context(track_event.guild_id)
-                                    .unwrap();
-
-                                if let Err(why) = player.finish(true) {
-                                    error!(
-                                        "Error sending finish message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
-                                }
-
-                                if let Err(why) = player.update_track(track_event.track.into()) {
-                                    error!(
-                                        "Error sending update track message for player {}: {}",
-                                        track_event.guild_id.0, why
-                                    );
-                                }
-
-                                ed.parse_and_dispatch(&x, |e| e.track_stuck).await
-                            }
+                            "TrackStuckEvent" => ed.parse_and_dispatch(&x, |e| e.track_stuck).await,
                             "WebSocketClosedEvent" => {
                                 ed.parse_and_dispatch(&x, |e| e.websocket_closed).await
                             }